# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ALERYAF HUB API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /warehouses:
    get:
      operationId: getWarehouses
      tags: [warehouses]
      summary: List all warehouses
      responses:
        "200":
          description: List of warehouses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Warehouse"
    post:
      operationId: createWarehouse
      tags: [warehouses]
      summary: Create a warehouse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWarehouseInput"
      responses:
        "201":
          description: Created warehouse
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warehouse"
  /warehouses/{id}:
    delete:
      operationId: deleteWarehouse
      tags: [warehouses]
      summary: Delete a warehouse
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /items:
    get:
      operationId: getItems
      tags: [items]
      summary: List all items
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
    post:
      operationId: createItem
      tags: [items]
      summary: Create an item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateItemInput"
      responses:
        "201":
          description: Created item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
  /items/{id}:
    put:
      operationId: updateItem
      tags: [items]
      summary: Update an item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateItemInput"
      responses:
        "200":
          description: Updated item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
    delete:
      operationId: deleteItem
      tags: [items]
      summary: Delete an item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /items/{id}/aliases:
    get:
      operationId: getItemAliases
      tags: [items]
      summary: Get aliases for an item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of aliases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ItemAlias"
    post:
      operationId: addItemAlias
      tags: [items]
      summary: Add alias for an item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateItemAliasInput"
      responses:
        "201":
          description: Created alias
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemAlias"
  /items/aliases/{aliasId}:
    delete:
      operationId: deleteItemAlias
      tags: [items]
      summary: Delete an item alias
      parameters:
        - name: aliasId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /profit/analysis:
    get:
      operationId: getProfitAnalysis
      tags: [profit]
      summary: Get profit analysis
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: currency
          in: query
          schema:
            type: string
            enum: [TRY, USD]
        - name: branchId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Profit analysis data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfitAnalysis"
  /profit/by-item:
    get:
      operationId: getProfitByItem
      tags: [profit]
      summary: Get profit breakdown by item
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: currency
          in: query
          schema:
            type: string
            enum: [TRY, USD]
        - name: branchId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Profit by item
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ItemProfit"
components:
  schemas:
...
    Warehouse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        branchId:
          type: integer
        branchName:
          type: string
        branchCode:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - branchId
        - branchName
        - isActive
        - createdAt
        - updatedAt
    CreateWarehouseInput:
      type: object
      properties:
        name:
          type: string
        branchId:
          type: integer
      required:
        - name
        - branchId
...
    Item:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
        nameAr:
          type: string
        category:
          type: string
        unitCostTry:
          type: number
        unitCostUsd:
          type: number
        unitPriceTry:
          type: number
        unitPriceUsd:
          type: number
        currentStock:
          type: number
        minStock:
          type: number
        isActive:
          type: boolean
        aliases:
          type: array
          items:
            $ref: "#/components/schemas/ItemAlias"
        createdAt:
          type: string
      required:
        - id
        - code
        - name
        - isActive
        - currentStock
        - createdAt
    CreateItemInput:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        nameAr:
          type: string
        category:
          type: string
        unitCostTry:
          type: number
        unitCostUsd:
          type: number
        unitPriceTry:
          type: number
        unitPriceUsd:
          type: number
        currentStock:
          type: number
        minStock:
          type: number
        isActive:
          type: boolean
      required:
        - code
        - name
    ItemAlias:
      type: object
      properties:
        id:
          type: integer
        itemId:
          type: integer
        alias:
          type: string
      required:
        - id
        - itemId
        - alias
    CreateItemAliasInput:
      type: object
      properties:
        alias:
          type: string
      required:
        - alias
...
    ProfitAnalysis:
      type: object
      properties:
        totalRevenue:
          type: number
        totalCost:
          type: number
        totalProfit:
          type: number
        profitMargin:
          type: number
        currency:
          type: string
        monthlyData:
          type: array
          items:
            $ref: "#/components/schemas/MonthlyProfitPoint"
      required:
        - totalRevenue
        - totalCost
        - totalProfit
        - profitMargin
        - currency
        - monthlyData
    MonthlyProfitPoint:
      type: object
      properties:
        month:
          type: string
        revenue:
          type: number
        cost:
          type: number
        profit:
          type: number
        margin:
          type: number
      required:
        - month
        - revenue
        - cost
        - profit
        - margin
    ItemProfit:
      type: object
      properties:
        itemId:
          type: integer
        itemCode:
          type: string
        itemName:
          type: string
        quantitySold:
          type: number
        revenue:
          type: number
        cost:
          type: number
        profit:
          type: number
        margin:
          type: number
      required:
        - itemId
        - itemCode
        - itemName
        - quantitySold
        - revenue
        - cost
        - profit
        - margin