# lib/api-zod/src/generated/types/profitAnalysis.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ALERYAF HUB API
 * OpenAPI spec version: 0.1.0
 */
import type { MonthlyProfitPoint } from "./monthlyProfitPoint";

export interface ProfitAnalysis {
  totalRevenue: number;
  totalCost: number;
  totalProfit: number;
  profitMargin: number;
  currency: string;
  monthlyData: MonthlyProfitPoint[];
}